refactor: 지난 주 로또 번호 parameter명 외 2
- answer -> winnigNumber
- static ZERO 변수명을 의미 있게 변경 ZERO -> PRIZE_MIN, MONEY_MIN
- 로또 프로그램에서 사용 되는 에러 throw LottoException

+++ src/main/java/lotto/view/ResultView.java
@@ -18,7 +18,7 @@ public class ResultView {
     private static final String PRINT_SUFFIX = "]";
 
 
-    public static void resultBuyCount(final long count) {
+    public static void resultBuyCount(final int count) {
         System.out.printf((BUY_COUNT_MESSAGE) + "%n", count);
     }
